# tests/contrato_openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginPeticion": {
        "properties": {
          "clave": {
            "title": "Clave",
            "type": "string"
          },
          "correo": {
            "title": "Correo",
            "type": "string"
          }
        },
        "required": [
          "correo",
          "clave"
        ],
        "title": "LoginPeticion",
        "type": "object"
      },
      "RefrescarTokenPeticion": {
        "properties": {
          "token_refresco": {
            "minLength": 20,
            "title": "Token Refresco",
            "type": "string"
          }
        },
        "required": [
          "token_refresco"
        ],
        "title": "RefrescarTokenPeticion",
        "type": "object"
      },
      "RegistroPeticion": {
        "properties": {
          "apellidos": {
            "maxLength": 100,
            "title": "Apellidos",
            "type": "string"
          },
          "clave": {
            "maxLength": 255,
            "minLength": 6,
            "title": "Clave",
            "type": "string"
          },
          "codigo_postal": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Postal"
          },
          "correo": {
            "format": "email",
            "title": "Correo",
            "type": "string"
          },
          "direccion": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion"
          },
          "distrito_id": {
            "anyOf": [
              {
                "maxLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distrito Id"
          },
          "nombre": {
            "maxLength": 100,
            "title": "Nombre",
            "type": "string"
          },
          "referencia": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referencia"
          },
          "telefono": {
            "anyOf": [
              {
                "maxLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "required": [
          "nombre",
          "apellidos",
          "correo",
          "clave"
        ],
        "title": "RegistroPeticion",
        "type": "object"
      },
...
      "RestablecerClavePeticion": {
        "properties": {
          "nueva_clave": {
            "maxLength": 255,
            "minLength": 6,
            "title": "Nueva Clave",
            "type": "string"
          },
          "token_recuperacion": {
            "minLength": 20,
            "title": "Token Recuperacion",
            "type": "string"
          }
        },
        "required": [
          "token_recuperacion",
          "nueva_clave"
        ],
        "title": "RestablecerClavePeticion",
        "type": "object"
      },
      "SolicitarRecuperacionClavePeticion": {
        "properties": {
          "correo": {
            "title": "Correo",
            "type": "string"
          }
        },
        "required": [
          "correo"
        ],
        "title": "SolicitarRecuperacionClavePeticion",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/autenticacion/iniciar-sesion"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Arquetipo Idiomatico Escalable de FastAPI",
    "title": "FastAPI Arquetipo",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/autenticacion/cerrar-sesion": {
      "post": {
        "operationId": "cerrar_sesion_api_autenticacion_cerrar_sesion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrescarTokenPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cerrar Sesion",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/cerrar-sesion-todos": {
      "post": {
        "operationId": "cerrar_sesion_todos_api_autenticacion_cerrar_sesion_todos_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Cerrar Sesion Todos",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/iniciar-sesion": {
      "post": {
        "operationId": "iniciar_sesion_api_autenticacion_iniciar_sesion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Iniciar Sesion",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/refrescar-token": {
      "post": {
        "operationId": "refrescar_token_api_autenticacion_refrescar_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefrescarTokenPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refrescar Token",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/registrar-cuenta": {
      "post": {
        "operationId": "registrar_cuenta_api_autenticacion_registrar_cuenta_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Registrar Cuenta",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/restablecer-clave": {
      "post": {
        "operationId": "restablecer_clave_api_autenticacion_restablecer_clave_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestablecerClavePeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restablecer Clave",
        "tags": [
          "Autenticacion"
        ]
      }
    },
    "/api/autenticacion/solicitar-recuperacion-clave": {
      "post": {
        "operationId": "solicitar_recuperacion_clave_api_autenticacion_solicitar_recuperacion_clave_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitarRecuperacionClavePeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Solicitar Recuperacion Clave",
        "tags": [
          "Autenticacion"
        ]
      }
    },
...
  }
}